e2e(redesign-plugin-menu): Use shared helper function for tests
