+++ e2e-tests/tests/423-text-plugin-list.ts
@@ -1,15 +1,9 @@
+import { addNewTextPlugin } from './helpers/add-plugin'
 import { popupWarningFix } from './helpers/popup-warning-fix'
 
 Feature('Serlo Editor - Text plugin - list')
 
 Before(popupWarningFix)
-
-function addNewTextPlugin(I) {
-  I.click('$add-new-plugin-row-button')
-  I.type('Text')
-  I.pressKey('Tab')
-  I.pressKey('Enter')
-}
 
 Scenario('Unordered list shortcuts', ({ I }) => {
   I.amOnPage('/entity/create/Article/1377')
