+++ e2e-tests/tests/helpers/add-plugin.ts
@@ -1,0 +1,8 @@
+import { selectors } from './selectors'
+
+export function addNewTextPlugin(I) {
+  I.click(selectors.addNewPluginButton)
+  I.type('Text')
+  I.pressKey('Tab')
+  I.pressKey('Enter')
+}
