+++ e2e-tests/tests/helpers/selectors.ts
@@ -1,0 +1,8 @@
+export const selectors = {
+  textEditor: '$plugin-text-editor',
+  addNewPluginButton: '$add-new-plugin-row-button',
+  multimediaWrapper: '$plugin-multimedia-wrapper',
+  toolbarUndoButton: '$editor-toolbar-undo',
+  toolbarRedoButton: '$editor-toolbar-redo',
+  numberedListButton: '$plugin-toolbar-button-nummerierte-liste',
+}
